Bug fix. When "theView" is a view that's embedded in multiple layers of superviews/controllers, the popup can appear in an incorrect position. In simpler view embeddings (such as common on iPhone) view.bounds and view.frame are often equal and so there's no issue.
NOTE: not 100% certain this is the only needed change. Possible same change needs to be made 2 lines up (line 203). I don't use code in a way that hits that execution path, so do not have any data to rely on. But I *am* 100% certain this line's change is needed based on functioning code in a live app.